% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marianapackage-package.R
\docType{package}
\name{marianapackage-package}
\alias{marianapackage}
\alias{marianapackage-package}
\title{marianapackage: A Toy Package for Illustration Purposes}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
