% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplier.R
\name{multiplier}
\alias{multiplier}
\title{A baby function to illustrate how to document stuff in R}
\arguments{
\item{a}{A real number}

\item{b}{A real number}
}
\value{
The product of \code{a} and \code{b}
}
\description{
A baby function to illustrate how to document stuff in R
}
\examples{
multiplyr(2, 5)
multiplyr(4, 8)
}
